# design/archive/DI_OPENAPI_v2.1.yaml
openapi: 3.1.0
info:
  title: Verigence Document Intelligence API
  version: 2.1.0
  description: 'Baselined Phase-1 vendor-neutral REST API.


    Verigence Document Intelligence is standalone and owns a lightweight Subject Registry.

    Primary business lookup is tenantId + subjectId. Machine lifecycle is Upload -> Process -> Confirm.

    Each successfully extracted document has confidenceScore 0-100.

    confidenceScore > 90.00 => humanVerificationStatus OPTIONAL; confidenceScore <= 90.00 => MANDATORY.

    verificationState separately records NOT_VERIFIED or VERIFIED.

    One vendor-neutral X-Correlation-ID is used end-to-end; no vendor tracing product or outbound event API is required
    in Phase 1.'
servers:
- url: https://{host}
  variables:
    host:
      default: document-intelligence.example
security:
- bearerAuth: []
...
paths:
...
  /v1/tenants/{tenantId}/subjects/{subjectId}/requirement-profile:
    put:
      tags:
      - Requirement Configuration
      summary: Assign one published Requirement Profile version to a Subject
      operationId: assignRequirementProfile
      parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/SubjectId'
      - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - requirementProfileId
              properties:
                requirementProfileId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Assignment active.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectRequirementAssignment'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /v1/tenants/{tenantId}/document-requirement-profiles:
    get:
      tags:
      - Requirement Configuration
      summary: List Requirement Profile versions
      operationId: listRequirementProfiles
      parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Profiles.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RequirementProfile'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
    post:
      tags:
      - Requirement Configuration
      summary: Create a DRAFT Requirement Profile version
      operationId: createRequirementProfile
      parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequirementProfileRequest'
      responses:
        '201':
          description: Draft created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementProfile'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /v1/tenants/{tenantId}/document-requirement-profiles/{profileId}/publish:
    post:
      tags:
      - Requirement Configuration
      summary: Publish an immutable Requirement Profile version
      operationId: publishRequirementProfile
      parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/ProfileId'
      - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Published profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementProfile'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      description: Publishes the DRAFT version and atomically retires the previously published version of the same
        profileKey, if any. Existing Subject assignments to an older immutable version remain valid until explicitly
        reassigned.
...
  /v1/tenants/{tenantId}/document-requirement-profiles/{profileId}:
    get:
      tags:
      - Requirement Profiles
      summary: Get one Requirement Profile version
      operationId: getRequirementProfile
      parameters:
      - $ref: '#/components/parameters/TenantId'
      - &id002
        in: path
        name: profileId
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Requirement Profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementProfile'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '404':
          $ref: '#/components/responses/Problem'
    put:
      tags:
      - Requirement Profiles
      summary: Replace the contents of a DRAFT Requirement Profile version
      operationId: updateDraftRequirementProfile
      parameters:
      - $ref: '#/components/parameters/TenantId'
      - *id002
      - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRequirementProfileDraftRequest'
      responses:
        '200':
          description: Updated DRAFT profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementProfile'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
      description: Updates only a DRAFT Requirement Profile. profileKey and version are immutable; published/retired
        versions cannot be edited.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    systemBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantId:
      in: path
      name: tenantId
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    SubjectId:
      in: path
      name: subjectId
      required: true
      schema:
        type: string
        format: uuid
...
    ProfileId:
      in: path
      name: profileId
      required: true
      schema:
        type: string
        format: uuid
...
    CorrelationId:
      in: header
      name: X-Correlation-ID
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: ^[A-Za-z0-9._:-]+$
      description: End-to-end execution correlation token. Verigence validates and propagates a supplied value;
        if absent it generates a UUID. The same value is returned in the response and propagated through the initial
        processing chain.
  schemas:
...
    RequirementProfileItemInput:
      type: object
      required:
      - documentTypeKey
      - requirementClassification
      properties:
        documentTypeKey:
          type: string
        requirementClassification:
          type: string
          enum:
          - MANDATORY
          - OPTIONAL
        minimumCount:
          type: integer
          minimum: 1
          default: 1
        displaySequence:
          type: integer
          default: 100
    CreateRequirementProfileRequest:
      type: object
      required:
      - profileKey
      - items
      properties:
        profileKey:
          type: string
          maxLength: 120
        description:
          type: string
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/RequirementProfileItemInput'
    RequirementProfile:
      type: object
      required:
      - requirementProfileId
      - profileKey
      - version
      - status
      - items
      properties:
        requirementProfileId:
          type: string
          format: uuid
        profileKey:
          type: string
        version:
          type: integer
        status:
          type: string
          enum:
          - DRAFT
          - PUBLISHED
          - RETIRED
        description:
          type:
          - string
          - 'null'
        items:
          type: array
          items:
            $ref: '#/components/schemas/RequirementProfileItemInput'
    RequirementProfileReference:
      type: object
      required:
      - requirementProfileId
      - profileKey
      - version
      properties:
        requirementProfileId:
          type: string
          format: uuid
        profileKey:
          type: string
        version:
          type: integer
    SubjectRequirementAssignment:
      type: object
      required:
      - tenantId
      - subjectId
      - requirementProfile
      properties:
        tenantId:
          type: string
        subjectId:
          type: string
          format: uuid
        requirementProfile:
          $ref: '#/components/schemas/RequirementProfileReference'
...
    Problem:
      type: object
      required:
      - status
      - code
      - title
      properties:
        status:
          type: integer
        code:
          type: string
          enum:
          - INVALID_REQUEST
          - UNAUTHORIZED
          - FORBIDDEN
          - TENANT_NOT_FOUND
          - TENANT_NOT_READY
          - RETENTION_POLICY_NOT_CONFIGURED
          - DEVICE_NOT_REGISTERED
          - REQUIREMENT_PROFILE_NOT_ASSIGNED
          - REQUIREMENT_PROFILE_NOT_PUBLISHED
          - DOCUMENT_TYPE_NOT_FOUND
          - EXTRACTION_PROFILE_NOT_FOUND
          - SUBJECT_DOCUMENT_NOT_FOUND
          - DOCUMENT_NOT_CONFIRMED
          - DOCUMENT_CONTENT_PURGED
          - IDEMPOTENCY_CONFLICT
          - FILE_EMPTY
          - FILE_TOO_LARGE
          - MIME_TYPE_NOT_ALLOWED
          - INVALID_FILE_CONTENT
          - INVALID_CONFIGURATION
          - DOCUMENT_NOT_FOUND
          - UNASSIGNED_DOCUMENT_NOT_FOUND
          - SUBJECT_DOCUMENT_MISMATCH
          - INVALID_DOCUMENT_STATE
          - DOCUMENT_REPLACEMENT_NOT_ALLOWED
          - DOCUMENT_ALREADY_ASSIGNED
          - PROFILE_IMMUTABLE
          - PROFILE_NOT_DRAFT
          - WHATSAPP_ROUTE_NOT_FOUND
          - QUARANTINE_ITEM_NOT_FOUND
          - STORAGE_WRITE_FAILED
          - STORAGE_READ_FAILED
          - QUALITY_POLICY_NOT_CONFIGURED
          - CLASSIFICATION_AMBIGUOUS
          - SUBJECT_MATCH_AMBIGUOUS
        title:
          type: string
        detail:
          type: string
        correlationId:
          type:
          - string
          - 'null'
          maxLength: 128
...
    UpdateRequirementProfileDraftRequest:
      type: object
      required:
      - items
      properties:
        description:
          type:
          - string
          - 'null'
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/RequirementProfileItemInput'
...
  responses:
    Problem:
      description: Standard problem response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
  headers:
    CorrelationId:
      description: End-to-end execution correlation identifier.
      schema:
        type: string
        minLength: 1
        maxLength: 128